All tables have the by fields
All tables will have inserted_by and updated_by. Some tables now have a
user_id field that seems to make inserted_by and updated_by irrelevant,
but the record could be attached to a particular user but
inserted/updated by a script. The script might have an associated
service account which will be the value of (inserted|updated)_by.