Every table that might be private will have an owner
Instead of some tables having users, and most tables having a private
field, where the private field would have to be tied to the inserted_by
field, we'll separate out the inserted_by and updated_by to be nothing
more than a record of who did what (inserted_by could be a script rather
than a user adding records). All tables will have an owned_by field and
a private field so that just about anything can be private.

This could cause problems later if multiple users want to add the same
novel, each privately, but we'll figure out the constraints.